MscmpSystHierarchy (mscmp_syst_hierarchy v0.1.0)

API providing a baseline methodology for defining Hierarchies and validating that implementations of those Hierarchies are valid.

The Hierarchy features supported by this Component are:

Providing a Hierarchy definition with which implementations can be validated.

Support for optionally validated Hierarchies ("structured" vs. "unstructured").

Support of optional Hierarchy levels.

Specification of which Hierarchy levels may be associated with "Leaf Nodes"; for example a Hierarchy defining an application menu structure may allow for levels which only define menu item groupings or may be associated directly with actionable application menu items where the menu items are "Leaf Nodes".

Protection of in-use Hierarchy definitions to ensure Hierarchy implementation operational consistency.

Important Note

While this Component makes certain guarantees about ensuring valid and consistent Hierarchy configurations, full consistency depends on Hierarchy implementing Components honoring the MscmpSystHierarchy guarantees. To point, Hierarchy implementations must implement database foreign keys to the Hierarchy defining tables and ensures their own data structures are able to be validated against the defined Hierarchies.

Concepts

There are a handful of conceptual ideas that you should be familiar with prior to working with this Component or one of the Hierarchy implementing Components.

Hierarchy

There are two different, but related, meanings assigned to the term "Hierarchy" in this Component.

In the first instance a Hierarchy is a configuration which establishes a pre- defined structure for data which can be represented in a tree-like structure. In this sense of the word we mean the totality of the records which define a distinct Hierarchy. This is the more generic and conceptual sense of the word and it will be the intended meaning in most contexts.

The second meaning of Hierarchy is a more technically focused term referring specifically to the Hierarchy "header record" as opposed to the Hierarchy Item records which are the "detail records" defining the levels of the Hierarchy. This second meaning will be meant more often when dealing with database and specific technical related concerns, including the specific configuration options which define the Hierarchy as a whole.

Hierarchies may be "structured" or "unstructured". Structured Hierarchies define specific levels of hierarchy which are required to some degree of Hierarchy implementing Component data. Unstructured Hierarchies aren't really Hierarchies at all, but allow implementing Components to optionally allow both structured and unstructured data without forcing special technical accommodations for the unstructured data.

Hierarchy Types

Hierarchies are assigned to a specific Hierarchy Type. The Hierarchy Type establishes the area of system functionality to which the Hierarchy sensibly applies. For example, a Hierarchy of "Type Menu" may define and be sensible to use in with application functions that define the application's user interface menus whereas a Hierarchy of "Type Product Category" would best be applied to Hierarchies which define product categorization structures.

Hierarchy Items

As mentioned above, Hierarchy Items specify the behaviors of the individual levels of hierarchy contained by the Hierarchy. This includes the configuration points which are applicable to individual levels of the Hierarchy. Hierarchy Item records are children of the Hierarchy records.

Hierarchy Items are required to be defined for all structured Hierarchies. Conversely, Hierarchy Items are prohibited in the context of unstructured Hierarchies.

Hierarchy Implementing Components

As mentioned above, MscmpSystHierarchy doesn't store or define the actual final data to which the Hierarchy is applied. This is the responsibility of other, specialized Components which conform to MscmpSystHierarchy implementation expectations.

Hierarchies

get_hierarchy_id_by_name(hierarchy_name)

@spec get_hierarchy_id_by_name(MscmpSystHierarchy.Types.hierarchy_name()) ::
  {:ok, MscmpSystHierarchy.Types.hierarchy_id()} | {:error, MscmpSystError.t()}

Returns the record ID of the Hierarchy as identified by its Internal Name.

On successful execution the Hierarchy record ID value is returned via a success tuple in the form {:ok, <Hierarchy ID>}. If an error is encountered or the requested record does not exist an error tuple in the form {:error, %MscmpSystError{}} is returned instead.

Parameters

hierarchy_name - the Internal Name of the desired Hierarchy record.

Examples

Normal call successfully returning a Hierarchy record ID value.

iex> {:ok, hierarchy_id} =
...>   MscmpSystHierarchy.get_hierarchy_id_by_name("hierarchy_example")
iex> is_binary(hierarchy_id)
true

A failing call due to requesting a non-existent Hierarchy record.

iex> {:error, %MscmpSystError{}} =
...>   MscmpSystHierarchy.get_hierarchy_id_by_name("nonexistent_hierarchy")

get_hierarchy_id_by_name!(hierarchy_name)

@spec get_hierarchy_id_by_name!(MscmpSystHierarchy.Types.hierarchy_name()) ::
  MscmpSystHierarchy.Types.hierarchy_id()

Returns the record ID of the Hierarchy as identified by its Internal Name, raising exceptions on error.

If the Hierarchy is not found or if some other error is encountered an exception is raised. This is same as get_hierarchy_id_by_name/1 with the exception of the error handling.

Parameters

hierarchy_name - the Internal Name of the desired Hierarchy record.

Examples

Normal call successfully returning a Hierarchy record ID value.

iex> hierarchy_id =
...>   MscmpSystHierarchy.get_hierarchy_id_by_name!("hierarchy_example")
iex> is_binary(hierarchy_id)
true

Enumeration Access

get_hierarchy_state_by_name(state_name)

@spec get_hierarchy_state_by_name(MscmpSystHierarchy.Types.hierarchy_state_name()) ::
  Msdata.SystEnumItems.t() | nil

Returns the Hierarchy State record for the given Hierarchy State Internal Name.

If the requested Hierarchy State record does not exist nil is returned instead. Any errors cause an exception to be raised.

Hierarchy States are currently implemented using the MscmpSystEnums Component.

Parameters

state_name - the Internal Name of the desired Hierarchy State to retrieve. This parameter is required.

Examples

Requesting an existing Hierarchy State:

iex> %Msdata.SystEnumItems{} =
...>   MscmpSystHierarchy.get_hierarchy_state_by_name("hierarchy_states_sysdef_active")

Requesting a non-existent Hierarchy State:

iex> MscmpSystHierarchy.get_hierarchy_state_by_name("nonexistent_state")
nil

get_hierarchy_state_default(functional_type \\ nil)

@spec get_hierarchy_state_default(
  MscmpSystHierarchy.Types.hierarchy_state_functional_types()
  | nil
) ::
  Msdata.SystEnumItems.t()

Returns the default Hierarchy State record.

Optionally, this function can also return the default Hierarchy State for a specified Hierarchy State Functional Type.

Any errors encountered will cause an exception to be raised.

Parameters

functional_type - an optional parameter specifying which Hierarchy State Functional Type's default record should be returned. If this parameter is nil (the parameter default), the default Hierarchy State overall will be returned.

Examples

Return the overall default Hierarchy State.

iex> %Msdata.SystEnumItems{internal_name: "hierarchy_states_sysdef_inactive"} =
...>   MscmpSystHierarchy.get_hierarchy_state_default()

Return the default Hierarchy State for the "active" Functional Type.

iex> %Msdata.SystEnumItems{internal_name: "hierarchy_states_sysdef_active"} =
...>   MscmpSystHierarchy.get_hierarchy_state_default(:hierarchy_states_active)

get_hierarchy_state_default_id(functional_type \\ nil)

@spec get_hierarchy_state_default_id(
  MscmpSystHierarchy.Types.hierarchy_state_functional_types()
  | nil
) ::
  MscmpSystHierarchy.Types.hierarchy_state_id()

Returns the default Hierarchy State record ID value.

Optionally, this function can also return the ID value of the default Hierarchy State for a specified Hierarchy State Functional Type.

Any errors encountered will cause an exception to be raised.

Parameters

functional_type - an optional parameter specifying which Hierarchy State Functional Type's default record ID should be returned. If this parameter is nil (the parameter default), the default Hierarchy State overall will be returned.

Examples

Return the overall default Hierarchy State.

iex> state_id = MscmpSystHierarchy.get_hierarchy_state_default_id()
iex> is_binary(state_id)
true

Return the default Hierarchy State for the "active" Functional Type.

iex> state_id = MscmpSystHierarchy.get_hierarchy_state_default_id(:hierarchy_states_active)
iex> is_binary(state_id)
true

get_hierarchy_state_id_by_name(state_name)

@spec get_hierarchy_state_id_by_name(MscmpSystHierarchy.Types.hierarchy_state_name()) ::
  MscmpSystHierarchy.Types.hierarchy_state_id() | nil

Returns the Hierarchy State ID for the given Hierarchy State Internal Name.

If the requested Hierarchy State record does not exist nil is returned instead. Any errors cause an exception to be raised.

Hierarchy States are currently implemented using the MscmpSystEnums Component.

Parameters

state_name - the Internal Name of the desired Hierarchy State ID to retrieve. This parameter is required.

Examples

Requesting an existing Hierarchy State:

iex> state_id =
...>   MscmpSystHierarchy.get_hierarchy_state_id_by_name("hierarchy_states_sysdef_inactive")
iex> is_binary(state_id)

Requesting a non-existent Hierarchy State:

iex> MscmpSystHierarchy.get_hierarchy_state_id_by_name("nonexistent_state")
nil

get_hierarchy_type_by_name(type_name)

@spec get_hierarchy_type_by_name(MscmpSystHierarchy.Types.hierarchy_type_name()) ::
  Msdata.SystEnumItems.t() | nil

Returns the Hierarchy Type record for the given Hierarchy Type Internal Name.

If the requested Hierarchy Type record does not exist nil is returned instead. Any errors cause an exception to be raised.

Hierarchy Types are currently implemented using the MscmpSystEnums Component.

Parameters

type_name - the Internal Name of the desired Hierarchy Type to retrieve. This parameter is required.

Examples

Requesting an existing Hierarchy Type:

iex> %Msdata.SystEnumItems{} =
...>   MscmpSystHierarchy.get_hierarchy_type_by_name("hierarchy_type_example")

Requesting a non-existent Hierarchy Type:

iex> MscmpSystHierarchy.get_hierarchy_type_by_name("nonexistent_type")
nil

get_hierarchy_type_id_by_name(type_name)

@spec get_hierarchy_type_id_by_name(MscmpSystHierarchy.Types.hierarchy_type_name()) ::
  MscmpSystHierarchy.Types.hierarchy_type_id() | nil

Returns the ID value of a Hierarchy Type record as identified by the Hierarchy Type's Internal Name.

If the requested Hierarchy Type record does not exist nil is returned instead. Any errors cause an exception to be raised.

Hierarchy Types are currently implemented using the MscmpSystEnums Component.

Parameters

type_name - the Internal Name of the desired Hierarchy Type ID to retrieve. This parameter is required.

Examples

Requesting an existing Hierarchy Type:

iex> id_value =
...>   MscmpSystHierarchy.get_hierarchy_type_id_by_name("hierarchy_type_example")
iex> is_binary(id_value)
true

Requesting a non-existent Hierarchy Type:

iex> MscmpSystHierarchy.get_hierarchy_type_id_by_name("nonexistent_type")
nil

list_hierarchy_types(opts \\ [])

@spec list_hierarchy_types(Keyword.t()) ::
  {:ok, [Msdata.SystEnumItems.t()]} | {:error, MscmpSystError.t()}

Lists the available Hierarchy Types, optionally sorted by the predefined sort order; doesn't raise exceptions on error.

Returns a list of Hierarchy Type records using a result tuple. If the request runs without error, a success tuple in the form of {:ok, <list>} is returned while an error condition will return an error tuple in the form of {:error, <error>}.

Parameters

opts - A Keyword List of optional parameters.

Options

:sorted (boolean/0) - When true, the hierarchy types will be returned in sorted order. The default value is true.

Examples

Retrieving the list of Hierarchy Types using the default options. Note that in the example there is only one available Hierarchy Type defined in the system.

iex> {:ok, [%Msdata.SystEnumItems{internal_name: "hierarchy_type_example"}]} =
...>   MscmpSystHierarchy.list_hierarchy_types()

list_hierarchy_types!(opts \\ [])

@spec list_hierarchy_types!(Keyword.t()) :: [Msdata.SystEnumItems.t()]

Lists the available Hierarchy Types, optionally sorted by the predefined sort order.

Returns a list of Hierarchy Type records. Raises an exception on errors. Other than the return behavior, this is the same as list_hierarchy_types/1.

Parameters

opts - A Keyword List of optional parameters.

Options

:sorted (boolean/0) - When true, the hierarchy types will be returned in sorted order. The default value is true.

Examples

Retrieving the list of Hierarchy Types using the default options. Note that in the example there is only one available Hierarchy Type defined in the system.

iex> [%Msdata.SystEnumItems{internal_name: "hierarchy_type_example"}] =
...>   MscmpSystHierarchy.list_hierarchy_types!()