Msdata.SystHierarchyItems (mscmp_syst_hierarchy v0.1.0)

Defines the hierarchy levels of a referenced Hierarchy. Modules implementing MscmpSystHierarchy into their data structures are expected to conform to the hierarchical structure defined by these records and test for conformance prior to allowing their data to be used in the application.

Defined in MscmpSystHierarchy.

Types

t()

@type t() :: %Msdata.SystHierarchyItems{
  __meta__: Ecto.Schema.Metadata.t(),
  allow_leaf_nodes: boolean() | nil,
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  display_name: String.t() | nil,
  external_name: String.t() | nil,
  hierarchy: term(),
  hierarchy_depth: integer() | nil,
  hierarchy_id: MscmpSystHierarchy.Types.hierarchy_id() | nil,
  id: MscmpSystHierarchy.Types.hierarchy_item_id() | nil,
  internal_name: MscmpSystHierarchy.Types.hierarchy_item_name() | nil,
  required: boolean() | nil
}